+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,39 +1,13 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import org.springframework.transaction.support.TransactionTemplate;
 
-public class UserService {
-    private final TransactionTemplate transactionTemplate;
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
+public interface UserService {
 
-    public UserService(final TransactionTemplate transactionTemplate,
-                       final UserDao userDao,
-                       final UserHistoryDao userHistoryDao) {
-        this.transactionTemplate = transactionTemplate;
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
+    User findById(final long id);
 
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
+    void insert(final User user);
 
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
+    void changePassword(final long id, final String newPassword, final String createBy);
 
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        transactionTemplate.execute(() -> {
-            final var user = findById(id);
-            user.changePassword(newPassword);
-            userDao.updatePassword(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-            return null;
-        });
-    }
 }
